Add FV test for Felix with k8s datastore driver
This commit adds and uses a separate FV test executable, whose code
lives under 'k8sfv'.  It tests Felix as a whole, configured to use
Kubernetes as its datastore; the system under test consists of

- Felix (configured to use k8s datastore)

- libcalico-go, including its k8s backend

- the Kubernetes API server (as provided by the
  gcr.io/google_containers/hyperkube-amd64 Docker image).

The test program drives that system by acting as a Kubernetes API
client, writing pod (and in future, namespace and network policy)
objects into the API server.

There are also new Makefile targets for running Prometheus and Grafana
containers to monitor the test.

Limit number of concurrent goroutines used for pod cleanup

Without this, in the cleanup for the 100k endpoint test, we get an error
from the API server saying that it has too much work and that we should
retry a delete operation later.