Allow selection of policy allow action
Allow selection of ACCEPT or RETURN for policy acceptance actions
to take when a security group is permitting the traffic. The
previous ACCEPT-only behavior did not allow for network-wide
policy to be inserted after the felix chains were checked.

RETURN will allow operators to filter traffic after it has passed
through the felix chains.