Add watching ippool
In the case of using gobgp as network backend, confd is only
used to execute allocate_ipip_addr command to sync ippool data.
So we can simplify calico-node if felix has process to watch ippool.
Then we will not need to run confd in the case of using gobgp.

This patch add the following functions.
- Add NetworkingBackend in config parameter.
- Watching ippool directory by thread, if any events occurs,
  it executes allocate_ipip_addr.

Signed-off-by: Fumihiko Kakuma <kakuma@valinux.co.jp>