Do not depend on the returned text string of iptables -w.
The iptables_lock_test is looking for a specific text string to be
returned by iptables. Unfortunately the output string varies depending
on version. Checking the returned status is sufficient for this test.
This problem was seen on a ppc64le. I verified the change with "make
fv" on both x86 and ppc64le.

Signed-off-by: David Wilder <wilder@us.ibm.com>