bpf: check source of tunnel pkts on dest node
The source node IP gets recorded when we first approve it and when we
first create a conntrack record. If the source node IP changed, we drop
such packets unless we are able to confirm that the node IP changed but
the node is the same. We cannot do that atm.