bpf: avoid need to accept_local on NodePort
All NATing is done on the first node (the one that forwards) instead of
doing the return SNAT when the packet is leaving the backing pod and is
being encaped.

The source IP of the encaped packet remains the client's IP and is fixed
up when leaving the forwarding node.

As a result, we never pass a packet with the local source IP through the
local IP stack when forwarding from ingress to egress, thus no need for
setting the accept_local=1.