bpf: nodeport return with multiple ifaces
In environments like EKS, where the egress route is determined by the
source IP as well, the return VXLAN packet for the nodeport tunnel may
end up at a wrong interface (and route) because the packet has the
source IP of the backing pod - to avoid being dropped by RPF check etc.

This fix uses the knowledge of the iface where the nodeport ingress
packets came from (recorded for the purpose of RPF check) and send the
packet out straight to the same interface.

Until we get bpf_redirect_neigh() in our supported kernels (5.10+) we
need to maintain an ARP lookup cache that allows us to patch in the
right MAC addresses so that the packet, once forwarded by the host
interface, can reach the next hop.

Each time we see an incoming VXLAN packet for the nodeport tunnel, we
record the IP it came from and the src/dst MAC addresses. We can use the
tunnel IP where we are sending the egress packet to look up the MAC
addresses we should patch in.

The lookup map is an LRU hash with 10k entries. That allows 10k other
nodes to forward their nodeports to a node.