bpf: fixup dst MAC for host networked nodeport backends
Because routing happens based on the dest IP (which is at that time the
client's) before we get hands on the packet on HEP, the MAC is likely
different (in the same subnet) than the MAC of the node that forwarded
the nodeport to us. Therefore once we place the packet in VXLAN and thus
change the dest IP, we must fix up the dest MAC based on what we
recorded when we received the inbound packets.

This is a similar case to the MAC fixing for regular pods. The only
difference is that the src MAC is already set correctly if we are on the
right device.

We cannot handle the case (yet) when the routes to the original client
and to the node on the other side of the vxland tunnel use different
NICs. We would need to redirect the packet to the other NIC first.