bpf: direct redirect in both directions
If one direction (HEP, with NAT) marks the flow for direct redirection
straight to the local workload, traffic from the workload is also
redirected straight to the iface the original traffic came from.

ARP needs to be recorded and MACs filled in as we bypass FIB and the
host net stack.

Traffic decaped from a vxlan tunnel records ARP befor ethe decap as we
would not have the right eth header after that.