bpf: direct redirection to WEP reuires ARP entries
Without having the right MAC addresses, net stack in the workload
rejects the packet as it is not considered for that host.

This needs to be programmed by felix.

The source MAC may not matter so much as workload can/will use ARP to
resolve the missing ARP entry using ARP procol between itself and the
host.