bpf: ToHostMark - EKS rpf check
env FELIX_BpfExtToServiceConnmark allows us to set a mark if a packet is
DNATed at the host interace (therefore is not comming from a pod), which
allows us to influence routing if the packet would go through the host
namespace - that means it is not forwarded to another interface
directly. We set a flag at the connection so that we mark all such
packets in both directions.

EKS uses a mark to trace nodeport connections to route them to correct
interface on the way back (the main ENI) in case the backend pod is
associated with a secondary ENI. Setting the mark allows us to follow
the same routing and pass RPF check.