bpf: ToHostMark - EKS rpf check
env FELIX_BPFTOHOSTMARK allows us to set a mark if a packet goes to host
namespace, that is, the packet is not forwarded directly in BPF. That
mark allows us to do extra stuff in the host namespace. For instance,
EKS uses a mark to trace nodeport connections to route them to correct
interface on the way back (the main ENI) in case the backend pod is
associated with a secondary ENI. Setting the mark allows us to follow
the same routing and pass RPF check.