bpf: source port collision detection
We already stop any packet if we detected that the source node
changed - part of our RPF check.

We now also stop any packet that would have a conflicting tracking
record after NAT. This means that it has the same 5-tuple after NAT,
which means that the source is the same and the destination is the
same but the destination was different before NAT.

The detection is race free in respect to any concurrent connection as we
first try to create the tracking entry atomically and if that fails, no
fwd entry it created.