bpf/proxy: exclude local workloads from NodePortRemotes
To work around the fact that CTLB does not resolve nodeports at the
node, but immediately at connect time, we need to know if there are any
local pods at the destination node if ExternalTrafficPolicy is set to
Local. Otherwise we would not be compliant with legacy kube-proxy if we
make connetion to a non-local pod of the destination node.

Note that CTLB would make such a connection with source preservation and
now extra hop.