bpf/proxy: fix conn cleaning when ExternalTrafficPolicy=Local
Our desire is to clean up connections from conntrack if we detect that a
given frontend no longer has a given backend. As a result, clients do
not need to wait for timeouts as the nodes would reply with RST for the
clients retransmits.

Because of wrong tracking which frontend may have which backend, when
ExternalTrafficPolicy=Local, we were incorrectly breaking perfectly good
connections.

This makes our decisions more conservative as we consider all currently
known backends as valid. There is no harm in not cleaning up connections
that couldn't have been started and letting them timeout if they exist.